Add support for user defined headers in client.
Gwobex functions can now use easily include a list
of headers in the the obex request. These will be sent
in the order they are specified. No checking for duplicate
headers is performed.

Incoming response headers are placed in a list in the order
of arrival in the gw_obex_xfer object, and so are accessible
to callback functions.

This also influences the order of sending headers - they are now
sent in the order: type, name, additional headers, body headers.