Add function to feed mimetype drivers with headers
This function calls the feed_next_header function of mimetype
drivers once for each header in the received object.
Afterwards, it calls feed_next_header with a OBEX_HDR_EMPTY
header to mark the end of received headers.

It is not possible to suspend from the feed_next_header function.