Support more fine-grained filtering of sdp records
Obex-client provides currently a method to filter sdp
records by service class. The service class is determined
from the Target dictionary entry passed to the CreateSession
method.

For some profiles it is usefull to provide better control
over the filtering process in order to weed out services
that do not provide the necessary functionality we need.
This patch provides this by adding an sdp_filter callback
to session_data. This method can be set based on the Target
dictionary entry.

This functionality in not strictly necessary, since one
can always specify the exact channel to use in an argument
to CreateSession, bypassing SDP querying entirely.

Additionally, a Parameters entry is added to the dictionary
passed to CreateSession. This is later passed to the sdp_filter
callback. A use case for this is for eg. to connect
to a service that has a specific Service ID, where the ID
is passed in string form through the Parameters entry.