gobex: add response function parameter to API
All *_req/*_req_pkt functions now have an additional parameter.
This parameter is the function to be used to handle incoming packets.
If this parameter is NULL, the default packet reponse function
is used.

Previously, the default response function set itself as the function
to be used to handle the next received packet. Now this is changed
so that the application supplied response function is called.